refactor(cli): split CLI parsing from main
Move clap command definitions and shorthand parsing into src/cli.rs so main.rs stays focused on process entry and command execution.

Closes #253
